capstone-project-comp3900-m18b-segmentationfault-main/back-end/requirements.txt
alembic==1.8.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
